Moves files into components folder. Test data made
