+++ Audio/data/output.csv
@@ -1,135 +1,549 @@
 note, length, volume
-0, 160, 0
-50, 65, 27
-0, 131, 4
-112, 67, 65
-78, 34, 0
-77, 35, 0
-78, 295, 34
-90, 34, 0
-89, 34, 0
-70, 33, 0
-77, 98, 29
-0, 134, 0
-84, 65, 32
-85, 265, 24
-87, 33, 0
-90, 33, 0
-91, 66, 64
-72, 32, 0
-93, 33, 0
-92, 31, 0
-73, 33, 0
-56, 33, 0
-63, 33, 0
-55, 32, 0
-72, 32, 0
-79, 68, 33
-91, 67, 29
-92, 33, 0
-79, 33, 0
-80, 32, 0
-92, 33, 0
-54, 32, 0
-80, 34, 0
-92, 65, 29
-0, 33, 0
-93, 66, 33
-94, 100, 23
-67, 38, 0
-58, 40, 0
-76, 41, 0
-83, 78, 31
-0, 120, 52
-94, 38, 0
-75, 114, 13
-82, 38, 0
-0, 76, 20
-94, 38, 0
-82, 117, 16
-56, 80, 35
-0, 39, 0
-52, 39, 0
-0, 103, 33
-94, 130, 2
-0, 134, 24
-69, 33, 0
-79, 34, 0
-72, 33, 0
-71, 67, 30
-59, 34, 0
-54, 32, 0
-50, 38, 0
-78, 68, 30
-77, 68, 33
-0, 34, 0
-92, 33, 0
-93, 35, 0
-92, 66, 40
-0, 34, 0
-112, 33, 0
-0, 32, 0
-92, 33, 0
-0, 67, 49
-92, 33, 0
-51, 66, 30
-92, 32, 0
-49, 33, 0
-0, 33, 0
-73, 65, 18
-0, 299, 15
-111, 33, 0
-0, 494, 15
-54, 34, 0
-0, 395, 3
-111, 32, 0
-112, 66, 2
-0, 163, 0
-110, 35, 0
-0, 327, 12
-110, 33, 0
-0, 97, 3
-105, 33, 0
-0, 98, 0
-51, 32, 0
-70, 34, 0
-0, 731, 11
-105, 33, 0
-0, 34, 0
-105, 34, 0
-86, 35, 0
-0, 99, 20
-50, 34, 0
-0, 138, 1
-49, 33, 0
-0, 439, 11
-65, 33, 0
-0, 66, 0
+0, 320, 0
+51, 212, 22
+52, 71, 16
+51, 106, 12
+52, 142, 22
+51, 70, 28
+50, 36, 0
+51, 315, 24
+50, 37, 0
+51, 34, 0
+0, 72, 0
+53, 181, 20
+54, 37, 0
+53, 73, 22
+51, 107, 18
 52, 34, 0
-0, 568, 7
-112, 34, 0
-0, 205, 32
-111, 33, 0
-0, 67, 46
-77, 34, 0
-0, 325, 43
-79, 34, 0
-0, 726, 1
-112, 33, 0
-0, 204, 10
-101, 33, 0
-102, 34, 0
+51, 70, 23
+52, 36, 0
+51, 104, 23
+0, 71, 9
+44, 36, 0
+46, 140, 7
+47, 107, 6
+46, 35, 0
+0, 71, 0
+52, 35, 0
+51, 425, 9
+50, 35, 0
+49, 35, 0
+46, 36, 0
+47, 35, 0
+48, 36, 0
+49, 179, 20
+48, 211, 9
+49, 70, 13
+50, 70, 0
+0, 243, 0
+53, 36, 0
+52, 181, 20
+51, 34, 0
+52, 70, 15
+51, 106, 13
+50, 36, 0
+0, 35, 0
+44, 37, 0
+45, 74, 10
+46, 255, 18
+47, 35, 0
+48, 36, 0
+49, 35, 0
+51, 317, 16
+52, 36, 0
+53, 35, 0
+55, 35, 0
+56, 106, 33
+55, 212, 20
+54, 71, 30
+53, 35, 0
+52, 35, 0
+51, 142, 23
+52, 348, 18
+51, 35, 0
+0, 139, 0
+53, 36, 0
+52, 487, 12
+0, 1036, 0
+59, 71, 39
+58, 254, 33
+57, 106, 41
+56, 71, 40
+55, 179, 33
+54, 35, 0
+53, 563, 32
+50, 70, 31
+51, 212, 29
+52, 35, 0
+53, 362, 26
+54, 37, 0
+53, 78, 10
+52, 37, 0
+48, 39, 0
+44, 37, 0
+45, 38, 0
+46, 118, 15
+47, 120, 6
+46, 39, 0
+45, 79, 8
+46, 37, 0
+45, 37, 0
+46, 38, 0
+48, 39, 0
+50, 38, 0
+51, 154, 16
+52, 35, 0
+51, 365, 16
+50, 35, 0
+47, 36, 0
+48, 36, 0
+47, 107, 28
+48, 69, 29
+47, 920, 29
+48, 35, 0
+47, 106, 28
+48, 70, 28
+47, 180, 26
+48, 35, 0
+47, 36, 0
+48, 34, 0
+47, 34, 0
+49, 32, 0
+0, 1789, 0
+49, 105, 24
+50, 36, 0
+51, 138, 16
+0, 36, 0
+53, 285, 18
+0, 424, 0
+45, 33, 0
+46, 34, 0
+47, 35, 0
+46, 35, 0
+0, 35, 0
+50, 35, 0
+49, 212, 19
+50, 459, 22
+0, 397, 0
+48, 35, 0
+47, 34, 0
+0, 70, 0
+51, 34, 0
+50, 179, 20
+0, 70, 0
+53, 36, 0
+52, 35, 0
+51, 282, 24
+52, 35, 0
+0, 424, 0
+45, 72, 20
+47, 34, 0
+48, 35, 0
+0, 141, 0
+52, 35, 0
+51, 34, 0
+50, 70, 17
+51, 178, 12
+52, 71, 17
+51, 70, 17
+52, 35, 0
+53, 69, 30
+52, 389, 25
+53, 35, 0
+0, 1405, 0
+53, 70, 26
+54, 35, 0
+53, 248, 23
+54, 106, 30
+53, 211, 23
+54, 35, 0
+53, 107, 17
+54, 33, 0
+53, 145, 15
+54, 31, 0
+53, 139, 14
+54, 35, 0
+0, 494, 0
+54, 36, 0
+55, 1513, 26
+0, 677, 0
+58, 35, 0
+57, 35, 0
+56, 1148, 27
+57, 69, 31
+58, 36, 0
+59, 74, 27
+58, 111, 24
+59, 73, 27
+58, 144, 22
+59, 106, 23
+58, 323, 22
+59, 36, 0
+58, 35, 0
+59, 35, 0
+58, 36, 0
+59, 35, 0
+58, 106, 21
+59, 35, 0
+58, 106, 17
+59, 70, 23
+58, 35, 0
+59, 36, 0
+0, 671, 1
+48, 35, 0
+0, 174, 0
+45, 36, 0
+0, 211, 3
+47, 285, 41
+46, 34, 0
+47, 37, 0
+46, 109, 34
+47, 36, 0
+46, 36, 0
+0, 109, 0
+68, 35, 0
+48, 180, 34
+50, 141, 27
+51, 139, 27
+52, 71, 33
+51, 72, 31
+52, 105, 34
+53, 106, 33
+54, 143, 36
+55, 211, 29
+56, 70, 28
+57, 141, 35
+56, 35, 0
+57, 140, 32
+58, 34, 0
+59, 177, 26
+58, 140, 23
+56, 105, 26
+55, 72, 20
+56, 106, 20
+55, 106, 0
+54, 35, 0
+52, 35, 0
+53, 249, 16
+54, 36, 0
+0, 649, 0
+60, 34, 0
+59, 70, 12
+60, 35, 0
+61, 35, 0
+60, 1830, 8
+59, 35, 0
+58, 1393, 10
+44, 37, 0
+43, 36, 0
+44, 35, 0
+56, 106, 21
+44, 106, 20
+56, 245, 19
+57, 36, 0
+56, 33, 0
+57, 34, 0
+56, 419, 19
+55, 528, 15
+0, 935, 0
+53, 318, 20
+54, 35, 0
+53, 31, 0
+54, 36, 0
+56, 36, 0
+55, 458, 21
+56, 427, 24
+0, 106, 0
+50, 106, 25
+51, 72, 21
+50, 35, 0
+51, 34, 0
+50, 177, 23
+51, 106, 19
+50, 106, 23
+51, 35, 0
+50, 106, 21
+51, 35, 0
+52, 35, 0
+51, 144, 19
+50, 34, 0
+49, 358, 27
+50, 147, 27
+49, 35, 0
+50, 283, 24
+49, 36, 0
+48, 34, 0
+49, 71, 25
+50, 35, 0
+49, 143, 28
+50, 35, 0
+49, 143, 25
+51, 35, 0
+52, 105, 13
+51, 351, 13
+0, 1902, 0
+52, 217, 18
+51, 36, 0
+50, 35, 0
 0, 33, 0
-77, 33, 0
-0, 465, 28
-98, 66, 51
-0, 372, 7
-69, 32, 0
-105, 32, 0
-0, 247, 21
-86, 35, 0
-105, 36, 0
-106, 34, 0
-107, 33, 0
-0, 895, 0
-110, 35, 0
+44, 35, 0
+45, 33, 0
+46, 34, 0
+47, 34, 0
+46, 106, 15
+47, 70, 16
+48, 35, 0
+0, 316, 0
+55, 71, 44
+54, 36, 0
+53, 178, 26
+52, 36, 0
+47, 34, 0
+46, 35, 0
+47, 36, 0
+48, 387, 21
+49, 70, 0
+0, 387, 0
+57, 37, 0
+56, 72, 29
+55, 105, 23
+56, 33, 0
+55, 73, 26
+52, 36, 0
+50, 185, 29
+51, 72, 24
+50, 111, 23
+51, 73, 0
+50, 36, 0
+0, 249, 0
+58, 35, 0
+57, 177, 31
+56, 142, 24
+55, 35, 0
+52, 36, 0
+53, 34, 0
+52, 249, 21
+51, 177, 13
+52, 70, 0
+0, 567, 0
+59, 34, 0
+58, 35, 0
+59, 70, 26
+58, 141, 22
+57, 35, 0
+51, 35, 0
+52, 33, 0
+53, 70, 34
+54, 36, 0
+53, 321, 20
+0, 217, 0
+59, 36, 0
+60, 146, 10
+61, 180, 10
+60, 69, 0
+59, 35, 0
+57, 35, 0
+55, 106, 0
+56, 71, 0
+55, 354, 0
+56, 35, 0
+55, 71, 0
+56, 35, 0
+55, 106, 0
+56, 248, 0
+55, 36, 0
+56, 35, 0
+55, 71, 0
+56, 213, 0
+55, 35, 0
+56, 175, 0
+0, 35, 0
+56, 70, 0
+0, 1077, 0
+45, 34, 0
+46, 70, 7
+47, 69, 11
+46, 597, 11
+47, 72, 10
+48, 35, 0
+49, 34, 0
+50, 69, 9
+51, 35, 0
+50, 36, 0
+51, 1242, 10
+50, 72, 16
+49, 35, 0
+47, 36, 0
+46, 36, 0
+44, 35, 0
+42, 35, 0
+41, 144, 2
+40, 36, 0
+0, 925, 0
+52, 40, 0
+51, 37, 0
+52, 38, 0
+51, 125, 17
+52, 39, 0
+51, 292, 11
+0, 212, 0
+53, 352, 24
+0, 177, 0
+56, 36, 0
+55, 356, 17
+0, 254, 0
+56, 402, 6
+0, 362, 0
+58, 108, 22
+57, 35, 0
+58, 2121, 11
+0, 437, 0
+60, 527, 7
+59, 70, 9
+58, 35, 0
+57, 72, 5
+58, 496, 11
+57, 105, 8
+56, 70, 5
+55, 35, 0
+56, 35, 0
+57, 71, 15
+56, 497, 1
+55, 423, 0
+54, 35, 0
+55, 142, 0
+0, 508, 0
+58, 1667, 14
+57, 35, 0
+54, 35, 0
+52, 36, 0
+53, 35, 0
+54, 35, 0
+53, 36, 0
+52, 103, 14
+53, 890, 17
+0, 470, 0
+56, 71, 21
+55, 209, 13
+54, 34, 0
+55, 566, 14
+54, 70, 5
+55, 69, 0
+0, 422, 0
+52, 105, 21
+53, 1468, 13
+52, 36, 0
+51, 35, 0
+50, 72, 15
+51, 1066, 8
+0, 1749, 0
+48, 107, 17
+49, 36, 0
+50, 71, 21
+51, 563, 15
+50, 35, 0
+0, 35, 0
+43, 36, 0
+44, 36, 0
+45, 36, 0
+46, 493, 18
+47, 35, 0
+48, 35, 0
+51, 36, 0
+52, 106, 11
+0, 70, 0
+53, 36, 0
+52, 106, 17
+53, 567, 11
+0, 618, 0
+54, 36, 0
+53, 638, 21
+52, 36, 0
+49, 34, 0
+0, 36, 0
+44, 35, 0
+45, 36, 0
+46, 283, 12
+45, 35, 0
+46, 177, 11
+45, 72, 7
+46, 35, 0
+0, 212, 0
+56, 71, 23
+55, 837, 14
+56, 38, 0
+0, 783, 0
+55, 36, 0
+54, 71, 32
+55, 566, 18
+54, 36, 0
+50, 36, 0
+47, 34, 0
+48, 36, 0
+49, 70, 24
+48, 528, 14
+49, 70, 0
+0, 70, 0
+56, 810, 18
+0, 839, 0
+56, 106, 30
+55, 72, 20
+56, 250, 22
+55, 178, 15
+54, 104, 23
+53, 572, 14
+54, 141, 11
+53, 106, 11
+54, 71, 18
+55, 35, 0
+57, 35, 0
+59, 71, 8
+58, 355, 11
+59, 34, 0
+58, 691, 5
+0, 634, 0
+59, 177, 20
+58, 71, 16
+59, 71, 16
+58, 386, 10
+59, 36, 0
+58, 249, 6
+59, 70, 12
+58, 245, 5
+59, 36, 0
+58, 70, 4
+59, 572, 8
+58, 37, 0
+59, 963, 8
+58, 957, 10
+0, 456, 0
+56, 1177, 24
+55, 775, 18
+56, 34, 0
+0, 460, 0
+54, 36, 0
+52, 35, 0
+53, 632, 20
+56, 35, 0
+55, 175, 20
+56, 398, 25
+55, 686, 18
+52, 34, 0
+0, 36, 0
+43, 35, 0
+42, 35, 0
+43, 35, 0
+45, 36, 0
+46, 36, 0
+45, 30, 0
+46, 36, 0
+47, 71, 25
+46, 1310, 22
+47, 35, 0
+48, 34, 0
+49, 34, 0
+50, 71, 19
+51, 1015, 19
+52, 70, 18
+51, 36, 0
+52, 71, 0
